# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from "./finding";

export interface AnalysisResult {
  reviewId: string;
  language: string;
  findings: Finding[];
  /** High-level summary of the analysis */
  summary: string;
  /** Code quality score from 0-100 */
  score: number;
  criticalCount: number;
  performanceCount: number;
  styleCount: number;
}