# lib/api-zod/src/generated/types/analyzeCodeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeCodeBody {
  /** The code snippet to analyze */
  code: string;
  /** Programming language (e.g. javascript, python, typescript) */
  language?: string;
  /** Optional GitHub URL to fetch code from */
  githubUrl?: string;
}