# lib/api-zod/src/generated/types/autofixCodeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from "./finding";

export interface AutofixCodeBody {
  /** The original code to fix */
  code: string;
  /** The findings to address */
  findings: Finding[];
  language?: string;
}