# lib/api-zod/src/generated/types/autofixResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutofixResult {
  /** The corrected code */
  fixedCode: string;
  /** Summary of what was changed */
  changesSummary: string;
  issuesFixed: number;
}