# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingCategory } from "./findingCategory";
import type { FindingSeverity } from "./findingSeverity";

export interface Finding {
  id: string;
  category: FindingCategory;
  severity: FindingSeverity;
  title: string;
  description: string;
  /** Line number where issue occurs (optional) */
  line?: number;
  /** Brief suggestion for fixing the issue */
  suggestion: string;
}