# lib/api-zod/src/generated/types/reviewStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewStats {
  totalReviews: number;
  totalFindings: number;
  criticalTotal: number;
  performanceTotal: number;
  styleTotal: number;
  averageScore: number;
  mostCommonIssue: string;
}