# lib/api-zod/src/generated/types/reviewSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewSummary {
  reviewId: string;
  language: string;
  score: number;
  criticalCount: number;
  performanceCount: number;
  styleCount: number;
  createdAt: Date;
  /** First 100 chars of the reviewed code */
  codeSnippet: string;
}